+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 - Initialized project
 - Added setup functions for setting up dev and prod logging
+
+## [0.0.2] - 2024-08-21
+
+- Added access counter for counting e.g. database access or outgoing API requests
+- Added detailed error structure